requirements.txt
python-dateutil==2.8.2
requests==2.31.0
flask==3.0.0
google-generativeai>=0.8.3
